Added new folder NavBar inside Components folder.NavBar refactorized and Logo and Button svg extracted into new componets
