+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
-import NavBar from "@/components/Navbar";
 import { ToastContainer } from 'react-toastify';
 import "./globals.css";
+import Navbar from "@/components/NavBar/Navbar";
 
 const poppins = Poppins({
   subsets: ["latin"],
@@ -24,7 +24,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={poppins.className}>
-        <NavBar />
+        <Navbar />
         <ToastContainer />
         <main className="max-w-7xl mx-auto py-24 px-4">{children}</main>
       </body>
